Fix: memo파일 수정
